move MD files to src_ions
