+++ ABACUS.develop/source/src_ions/MD_thermo.h
@@ -1,6 +1,7 @@
 #ifndef MD_THERMO_H
 #define MD_THERMO_H
-#include "tools.h"
+
+#include "src_pw/tools.h"
 
 class MD_thermo 
 {
